(cleanup) Move Splash.js to containers
